pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "advent-of-code-2022"
version = "0.1.0"
description = ""
authors = ["Tyson McNulty <mcnultyt@vmware.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
